refactor: avoid duplicate mappings (#131)
