+++ environment-dev.yml
@@ -1,11 +1,24 @@
 name: htsinfer
 channels:
-  - defaults
+  - conda-forge
+  - bioconda
 dependencies:
-  - coverage>=5.3
-  - flake8>=3.8.4
-  - mypy>=0.782
-  - pylint>=2.4.4
-  - pytest>=6.1.0
+  - biopython >=1.78
+  - coverage >=5.3
+  - cutadapt >=3.5, <=4.2
+  - flake8 >=3.8.4
+  - kallisto >=0.46.1, <= 0.48.0
+  - mypy >=0.782
+  - numpy >=1.22, <1.25
+  - pandas >=1.3.5, <1.4.0
+  - pip >=20.2.3
+  - pyahocorasick >=1.4.0
+  - pydantic >=1.8.1, <2
+  - pylint >=2.4.4
+  - pysam >=0.16.0
+  - pytest >=6.1.0
+  - python >=3.8, <=3.10
+  - star >=2.7.6
   - pip:
     - python-semantic-release>=7.15.0
+    - -e .
