+++ environment.yml
@@ -1,18 +1,18 @@
 name: htsinfer
 channels:
-  - defaults
-  - bioconda
   - conda-forge
+  - bioconda
 dependencies:
-  - biopython>=1.78
-  - kallisto>=0.46.1, <= 0.48.0
-  - pandas>=1.0.5
-  - pip>=20.2.3
-  - pyahocorasick>=1.4.0
-  - pydantic>=1.8.1, <2
-  - pysam>=0.16.0
-  - python>3.6, <3.10
-  - star>=2.7.6
-  - cutadapt>=3.5, <=4.2
+  - biopython >=1.78
+  - cutadapt >=3.5, <=4.2
+  - kallisto >=0.46.1, <= 0.48.0
+  - numpy >=1.22, <1.25
+  - pandas >=1.3.5, <1.4.0
+  - pip >=20.2.3
+  - pyahocorasick >=1.4.0
+  - pydantic >=1.8.1, <2
+  - pysam >=0.16.0
+  - python >=3.8, <=3.10
+  - star >=2.7.6
   - pip:
     - -e .
