+++ htsinfer/htsinfer.py
@@ -27,6 +27,7 @@ from htsinfer.models import (
     Config,
 )
 from htsinfer.subset_fastq import SubsetFastq
+from htsinfer.mapping import Mapping
 
 LOGGER = logging.getLogger(__name__)
 
@@ -64,6 +65,7 @@ class HtsInfer:
             else config.args.tmp_dir / config.args.transcripts_file.name
         )
         self.state: RunStates = RunStates.OKAY
+        self.mapping: Mapping = Mapping(config=self.config)
 
     def evaluate(self):
         """Determine library metadata."""
@@ -247,6 +249,7 @@ class HtsInfer:
         """Determine library type."""
         get_lib_type = GetLibType(
             config=self.config,
+            mapping=self.mapping,
         )
         get_lib_type.evaluate()
         self.config.results.library_type = get_lib_type.results
@@ -255,6 +258,7 @@ class HtsInfer:
         """Determine read orientation."""
         get_read_orientation = GetOrientation(
             config=self.config,
+            mapping=self.mapping,
         )
         self.config.results.read_orientation = get_read_orientation.evaluate()
