+++ tests/test_get_read_orientation.py
@@ -4,8 +4,6 @@ import pytest
 
 from htsinfer.exceptions import (
     FileProblem,
-    SamFileProblem,
-    StarProblem,
 )
 from htsinfer.get_read_orientation import GetOrientation
 from htsinfer.models import (
@@ -18,11 +16,8 @@ from htsinfer.models import (
     StatesTypeRelationship,
 )
 from tests.utils import (
-    FILE_2000_RECORDS,
-    FILE_DUMMY,
     FILE_EMPTY_ALIGNED_SAM,
     FILE_BAD_ALIGNED_SAM,
-    FILE_INVALID_TRANSCRIPTS,
     FILE_MATE_1,
     FILE_MATE_2,
     FILE_ORIENTATION_ISF_1,
@@ -38,11 +33,9 @@ from tests.utils import (
     FILE_UNMAPPED_PAIRED_1,
     FILE_UNMAPPED_PAIRED_2,
     FILE_UNMAPPED_SINGLE,
+    FILE_IDS_NOT_MATCH_2,
     CONFIG,
-    RaiseError,
-    SubprocessError,
-    SOURCE_HUMAN,
-    SOURCE_FRUIT_FLY,
+    MAPPING,
 )
 
 
@@ -54,7 +47,10 @@ class TestGetOrientation:
         CONFIG.args.path_1_processed = FILE_MATE_1
         CONFIG.args.path_2_processed = None
         CONFIG.args.t_file_processed = FILE_TRANSCRIPTS
-        test_instance = GetOrientation(config=CONFIG)
+        CONFIG.results.library_type = ResultsType()
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         assert test_instance.paths[0] == FILE_MATE_1
         assert test_instance.library_type == ResultsType()
         assert test_instance.transcripts_file == FILE_TRANSCRIPTS
@@ -64,7 +60,8 @@ class TestGetOrientation:
         CONFIG.args.path_1_processed = FILE_MATE_1
         CONFIG.args.path_2_processed = FILE_MATE_2
         CONFIG.args.t_file_processed = FILE_TRANSCRIPTS
-        test_instance = GetOrientation(config=CONFIG)
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         assert test_instance.paths[0] == FILE_MATE_1
         assert test_instance.paths[1] == FILE_MATE_2
         assert test_instance.library_type == ResultsType()
@@ -76,14 +73,14 @@ class TestGetOrientation:
         CONFIG.args.path_2_processed = FILE_MATE_2
         CONFIG.args.t_file_processed = FILE_TRANSCRIPTS
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         assert test_instance.paths[0] == FILE_MATE_1
         assert test_instance.paths[1] == FILE_MATE_2
         assert test_instance.library_type == ResultsType()
         assert test_instance.library_source == ResultsSource()
         assert test_instance.transcripts_file == FILE_TRANSCRIPTS
         assert test_instance.tmp_dir == tmpdir
-        assert test_instance.threads_star == 1
         assert test_instance.min_mapped_reads == 18
         assert test_instance.min_fraction == 0.75
 
@@ -94,7 +91,9 @@ class TestGetOrientation:
         CONFIG.args.path_1_processed = FILE_UNMAPPED_SINGLE
         CONFIG.args.path_2_processed = None
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.not_available,
@@ -110,7 +109,9 @@ class TestGetOrientation:
                 file_2=Source(),
             )
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.stranded_forward,
@@ -122,7 +123,9 @@ class TestGetOrientation:
         """Get read orientation for a single-end stranded reverse library."""
         CONFIG.args.path_1_processed = FILE_ORIENTATION_SR
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.stranded_reverse,
@@ -134,7 +137,9 @@ class TestGetOrientation:
         """Get read orientation for a single-end unstranded library."""
         CONFIG.args.path_1_processed = FILE_ORIENTATION_U
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.unstranded,
@@ -153,7 +158,9 @@ class TestGetOrientation:
         CONFIG.results.library_type = ResultsType(
             relationship=StatesTypeRelationship.split_mates,
         )
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.not_available,
@@ -174,7 +181,9 @@ class TestGetOrientation:
         )
         CONFIG.args.t_file_processed = FILE_TRANSCRIPTS
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.stranded_forward,
@@ -187,7 +196,9 @@ class TestGetOrientation:
         CONFIG.args.path_1_processed = FILE_ORIENTATION_ISR_1
         CONFIG.args.path_2_processed = FILE_ORIENTATION_ISR_2
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.stranded_reverse,
@@ -200,7 +211,9 @@ class TestGetOrientation:
         CONFIG.args.path_1_processed = FILE_ORIENTATION_IU_1
         CONFIG.args.path_2_processed = FILE_ORIENTATION_IU_2
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.unstranded,
@@ -208,183 +221,21 @@ class TestGetOrientation:
             relationship=StatesOrientationRelationship.inward_unstranded,
         )
 
-    def test_subset_transcripts_by_organism(self, tmpdir):
-        """Get filtered orgainsm transcripts for different organisms."""
-        CONFIG.results.library_type = ResultsType(
-                relationship=StatesTypeRelationship.split_mates,
-            )
-        CONFIG.results.library_source = ResultsSource(
-                file_1=SOURCE_HUMAN,
-                file_2=SOURCE_FRUIT_FLY
-            )
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        results = test_instance.subset_transcripts_by_organism()
-        filtered_organisms_transcripts = \
-            tmpdir / f"{CONFIG.results.library_source}.fasta"
-        assert results == filtered_organisms_transcripts
-
-    def test_subset_transcripts_by_organism_file_problem(self, tmpdir):
-        """Pass dummy file as transcripts.fasta file to simulate
-        file problem."""
-        CONFIG.args.path_2_processed = None
-        CONFIG.results.library_type = ResultsType()
-        CONFIG.results.library_source = ResultsSource()
-        CONFIG.args.t_file_processed = FILE_DUMMY
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-
-        with pytest.raises(FileProblem):
-            test_instance.subset_transcripts_by_organism()
-
-    def test_subset_transcripts_by_organism_invalid_fasta(self, tmpdir):
-        """Pass invalid transcripts.fasta file to simulate index error."""
-        CONFIG.results.library_source = ResultsSource(
-                file_1=SOURCE_HUMAN,
-                file_2=SOURCE_FRUIT_FLY
-        )
-        CONFIG.args.t_file_processed = FILE_INVALID_TRANSCRIPTS
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        results = test_instance.subset_transcripts_by_organism()
-        filtered_organisms_transcripts = \
-            tmpdir / f"{CONFIG.results.library_source}.fasta"
-        assert results == filtered_organisms_transcripts
-
-    def test_get_fasta_size(self, tmpdir):
-        """Get nucleotide statistics for filtererd transcripts
-        with different organisms."""
-        CONFIG.results.library_source = ResultsSource(
-                file_1=SOURCE_HUMAN,
-                file_2=SOURCE_FRUIT_FLY,
-            )
-        CONFIG.args.path_2_processed = FILE_ORIENTATION_IU_2
-        CONFIG.results.library_type = ResultsType(
-            relationship=StatesTypeRelationship.split_mates,
-        )
-        CONFIG.args.t_file_processed = FILE_TRANSCRIPTS
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        filtered_organisms_transcripts = \
-            test_instance.subset_transcripts_by_organism()
-        results = test_instance.get_fasta_size(filtered_organisms_transcripts)
-        assert results == 249986
-
-    def test_get_fasta_size_file_problem(self, tmpdir):
-        """Pass dummy file as filtered_organisms_transcripts
-        to simulate file problem."""
-        CONFIG.args.path_2_processed = None
-        CONFIG.results.library_type = ResultsType()
-        CONFIG.results.library_source = ResultsSource()
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        with pytest.raises(FileProblem):
-            test_instance.get_fasta_size(FILE_DUMMY)
-
-    def test_get_star_index_string_size(self, tmpdir):
-        """Get length of STAR SA pre-indexing string."""
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        results = test_instance.get_star_index_string_size(249986)
-        assert results == 7
-
-    def test_evaluate_star_index_problem(self, monkeypatch, tmpdir):
-        """Force raising exception to stimulate a star problem."""
-        CONFIG.results.library_source = ResultsSource(
-                file_1=SOURCE_HUMAN,
-                file_2=SOURCE_FRUIT_FLY,
-            )
-        CONFIG.args.path_2_processed = FILE_ORIENTATION_IU_2
-        CONFIG.results.library_type = ResultsType(
-                relationship=StatesTypeRelationship.split_mates,
-            )
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        monkeypatch.setattr(
-            'htsinfer.get_read_orientation.GetOrientation.create_star_index',
-            lambda *args, **kwargs: StarProblem
-        )
-        with pytest.raises(StarProblem):
-            test_instance.evaluate()
-
-    def test_prepare_star_alignment_commands(self, tmpdir):
-        """Get star alignment command."""
-        CONFIG.args.path_1_processed = FILE_2000_RECORDS
-        CONFIG.args.path_2_processed = None
-        CONFIG.results.library_type = ResultsType(
-            relationship=StatesTypeRelationship.not_mates,
-        )
-        CONFIG.results.library_source = ResultsSource(
-                file_1=SOURCE_HUMAN,
-                file_2=Source(),
-            )
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        index_dir = tmpdir / 'index'
-        file1_alignment_path = tmpdir / 'alignments/file_1'
-        cmd = "STAR --alignIntronMax 1 --alignEndsType Local --runThreadN 1" \
-            + " --genomeDir " + str(index_dir) + " --outFilterMultimapNmax " \
-            + "50 --outSAMunmapped Within KeepPairs --readFilesIn " \
-            + str(FILE_2000_RECORDS) + " --outFileNamePrefix " \
-            + str(file1_alignment_path) + "/"
-        results = test_instance.prepare_star_alignment_commands(
-            index_dir=index_dir
-            )
-        assert ' '.join(list(results.values())[0]) == cmd
-
-    def test_generate_star_alignments_problem(self, monkeypatch, tmpdir):
-        """Force raising exception to simulate problem."""
-        CONFIG.results.library_source = ResultsSource(
-                file_1=SOURCE_HUMAN,
-                file_2=SOURCE_FRUIT_FLY,
-            )
-        CONFIG.args.path_1_processed = FILE_ORIENTATION_IU_1
-        CONFIG.args.path_2_processed = FILE_ORIENTATION_IU_2
-        CONFIG.results.library_type = ResultsType(
-            relationship=StatesTypeRelationship.not_mates,
-        )
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        sub_method_name = 'htsinfer.get_read_orientation.' + \
-            'GetOrientation.generate_star_alignments'
-        monkeypatch.setattr(
-            sub_method_name,
-            lambda *args, **kwargs: StarProblem,
-        )
-        with pytest.raises(FileProblem):
-            test_instance.evaluate()
-
-    def test_generate_star_alignments_dummy_cmd(self, tmpdir):
-        """Pass dummy cmd to force simulate star problem."""
-        CONFIG.args.path_2_processed = None
-        CONFIG.results.library_type = ResultsType()
-        CONFIG.results.library_source = ResultsSource()
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        index_dir = tmpdir / 'index'
-        file1_alignment_path = tmpdir / 'alignments/file_1'
-        dummy_cmd = [
-            'STAR', '--alignIntrnMax', '1',
-            '--alignEndsType', 'Local', '--runThreadN', '1',
-            "--genomeDir", f"{str(index_dir)}",
-            ]
-        cmds = {file1_alignment_path: dummy_cmd}
-        with pytest.raises(StarProblem):
-            test_instance.generate_star_alignments(cmds)
-
-    def test_process_single_dummy_sam_file(self, tmpdir):
+    def test_process_single_dummy_sam_file_file_problem(self, tmpdir):
         """Pass dummy aligned.out.sam file to simulate file
         problem."""
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        with pytest.raises(SamFileProblem):
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
+        with pytest.raises(FileProblem):
             test_instance.process_single(FILE_EMPTY_ALIGNED_SAM)
 
     def test_process_paired_dummy_sam_file(self, tmpdir):
         """Pass dummy aligned.out.sam file to simulate file
         problem."""
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         with pytest.raises(FileProblem):
             test_instance.process_paired(FILE_EMPTY_ALIGNED_SAM)
 
@@ -392,69 +243,31 @@ class TestGetOrientation:
         """Pass bad_aligned.out.sam file to ensure correct
         paired file behaviour."""
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         assert test_instance.process_paired(FILE_BAD_ALIGNED_SAM) \
                == ResultsOrientation()
-
-    def test_create_star_index_star_problem(self, tmpdir):
-        """Pass invalid transcripts path to simulate star problem."""
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        transcripts_path = tmpdir / 'invalid'
-        with pytest.raises(StarProblem):
-            test_instance.create_star_index(transcripts_path)
 
     def test_evaluate_paired_not_mates_unmapped(self, tmpdir):
         """Get read orientation for a paired-end library with no mappable
         reads.
         """
         CONFIG.args.path_1_processed = FILE_UNMAPPED_PAIRED_1
-        CONFIG.args.path_2_processed = FILE_UNMAPPED_PAIRED_2
+        CONFIG.args.path_2_processed = FILE_IDS_NOT_MATCH_2
         CONFIG.results.library_type = ResultsType(
-                relationship=StatesTypeRelationship.not_mates,
+                relationship=StatesTypeRelationship.not_available,
             )
+        CONFIG.results.library_source = ResultsSource(
+            file_1=Source(),
+            file_2=Source(),
+        )
         CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
+        MAPPING.mapped = False
+        test_instance = GetOrientation(config=CONFIG,
+                                       mapping=MAPPING)
         results = test_instance.evaluate()
         assert results == ResultsOrientation(
             file_1=StatesOrientation.not_available,
-            file_2=StatesOrientation.not_available,
+            file_2=StatesOrientation.stranded_reverse,
             relationship=StatesOrientationRelationship.not_available,
         )
-
-    def test_subset_transcripts_by_organism_cannot_write_file(
-        self, monkeypatch, tmpdir
-    ):
-        """Force raising of ``OSError`` to simulate file problem."""
-        CONFIG.args.path_1_processed = FILE_ORIENTATION_IU_1
-        CONFIG.args.path_2_processed = None
-        CONFIG.results.library_source = ResultsSource()
-        CONFIG.results.library_type = ResultsType()
-        CONFIG.args.t_file_processed = FILE_INVALID_TRANSCRIPTS
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        monkeypatch.setattr(
-            'Bio.SeqIO.write',
-            RaiseError(exc=OSError),
-        )
-        with pytest.raises(FileProblem):
-            test_instance.subset_transcripts_by_organism()
-
-    def test_generate_star_alignments_star_problem(self, monkeypatch, tmpdir):
-        """Force raising of ``SubprocessError`` to simulate star probelm."""
-        CONFIG.args.t_file_processed = FILE_TRANSCRIPTS
-        CONFIG.args.tmp_dir = tmpdir
-        test_instance = GetOrientation(config=CONFIG)
-        file1_alignment_path = tmpdir / 'alignments/file_1'
-        dummy_cmd = [
-            'STAR', '--alignIntrnMax', '1',
-            '--alignEndsType', 'Local', '--runThreadN', '1',
-            "--genomeDir",
-            ]
-        cmds = {file1_alignment_path: dummy_cmd}
-        monkeypatch.setattr(
-            'subprocess.run',
-            lambda *args, **kwargs: SubprocessError(),
-        )
-        with pytest.raises(StarProblem):
-            test_instance.generate_star_alignments(cmds)
